fix run shell

+++ run_wsl.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env bash
+#!/usr/bin/env sh
 
 set -e
 
@@ -9,17 +9,23 @@ fi
 
 # shellcheck disable=SC1035
 if !(pip3 > /dev/null 2>&1); then
-  echo "start install pip..."
-  curl https://bootstrap.pypa.io/get-pip.py -o get-pip.py
-  python3 get-pip.py
-  # shellcheck disable=SC2016
-  echo 'export PATH="$HOME/.local/bin:$PATH"' >> ~/.bash_profile
-  source ~/.bash_profile
+  if [ ! -e get-pip.py ]; then
+    echo "startin' enable pip..."
+    curl https://bootstrap.pypa.io/get-pip.py -o get-pip.py
+    python3 get-pip.py
+    echo "complete."
+  fi
+  . ~/.profile
+fi
+
+python3 -m venv --without-pip venv
+. venv/bin/activate
+# shellcheck disable=SC2144
+if [ ! -e venv/lib/python*/site-packages/pdfminer ]; then
+  echo "startin' pip install..."
+  pip3 install --target venv/lib/python*/site-packages -r requirements.txt 1> /dev/null
   echo "complete."
 fi
 
-python3 -m venv --without-pip venv 1> /dev/null
-source venv/bin/activate 1> /dev/null
-pip3 install -r requirements.txt 1> /dev/null
 python3 __main__.py "$1"
 deactivate
